+++ src/gql/arguments/elements/Product.php
@@ -9,6 +9,7 @@ namespace craft\commerce\gql\arguments\elements;
 
 use Craft;
 use craft\commerce\elements\Product as ProductElement;
+use craft\commerce\gql\types\input\Variant;
 use craft\commerce\Plugin;
 use craft\gql\base\ElementArguments;
 use craft\gql\types\QueryArgument;
@@ -53,6 +54,11 @@ class Product extends ElementArguments
                 'type' => Type::listOf(QueryArgument::getType()),
                 'description' => 'Narrows the query results based on the product types the products belong to, per the product type IDs.'
             ],
+            'hasVariant' => [
+                'name' => 'hasVariant',
+                'type' => Variant::getType(),
+                'description' => 'Narrows the query results to only products that have certain variants.'
+            ]
         ]);
     }