+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 ### Added
 - Added `craft\commerce\events\ModifyCartInfoEvent::$cart`. ([#1536](https://github.com/craftcms/commerce/issues/1536))
+- Added the `hasVariant` argument to GraphQL product queries. ([#1544](https://github.com/craftcms/commerce/issues/1544))
+- Added `craft\commerce\gql\types\input\Variant`.
 
 ### Fixed
 - Fixed a bug products were incorrectly showing as having sales when using the `hasSales` query parameter.
+- Fixed a bug where the Edit Order page would save the previous order status message after only changing the status. ([#1366](https://github.com/craftcms/commerce/issues/1366))
 - Fixed a bug where it wasn’t possible to delete all line item statuses.
 
 ## 3.1.10 - 2020-06-23